requirements.txt
openai>=1.0.0
langchain>=0.2.0
langchain-community>=0.2.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
transformers>=4.40.0
datasets>=2.19.0
peft>=0.10.0
accelerate>=0.29.0
streamlit>=1.35.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
